requirements.txt
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
mysql-connector-python==8.0.29
protobuf==4.21.2
pycodestyle==2.8.0
pytz==2022.1
requests==2.28.1
toml==0.10.2
urllib3==1.26.10
websocket-client==1.3.3
zope.interface==5.4.0
